poly-sqlalchemy/requirements.txt
factory-boy==3.2.1
Faker==18.7.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
